fix: enum to uint32_t conversion

+++ Engine/Rendering/Private/Raycaster/VoxelPipeline.cpp
@@ -263,7 +263,7 @@ UploadDescriptor VoxelPipeline::GetUniformUploadDescriptor( const shared_ptr<GPU
     VkWriteDescriptorSet write = {
         .sType           = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET,
         .dstSet          = GetDescriptorSet(),
-        .dstBinding      = sr,
+        .dstBinding      = static_cast<uint32_t>( sr ),
         .descriptorCount = 1,
         .descriptorType  = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER,
         .pBufferInfo     = &bufferInfo,
